requirements.txt
amqp==2.3.2
Babel==2.6.0
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
debtcollector==1.20.0
dnspython==1.15.0
eventlet==0.24.1
extras==1.0.0
fasteners==0.14.1
fixtures==3.0.0
futurist==1.8.0
greenlet==0.4.15
idna==2.7
iso8601==0.1.12
Jinja2==2.10
kombu==4.2.2
linecache2==1.0.0
MarkupSafe==1.1.0
monotonic==1.5
msgpack==0.6.0
netaddr==0.7.19
netifaces==0.10.7
oslo.concurrency==3.29.0
oslo.config==6.7.0
oslo.context==2.22.0
oslo.i18n==3.23.0
oslo.log==3.42.0
oslo.messaging==9.2.1
oslo.middleware==3.37.0
oslo.serialization==2.28.1
oslo.service==1.33.0
oslo.utils==3.38.0
Paste==3.0.5
PasteDeploy==2.0.1
pbr==5.1.1
pkg-resources==0.0.0
prettytable==0.7.2
pyinotify==0.9.6
pyparsing==2.3.0
python-dateutil==2.7.5
python-mimeparse==1.6.0
pytz==2018.7
PyYAML==3.13
repoze.lru==0.7
requests==2.20.1
rfc3986==1.2.0
Routes==2.4.1
six==1.11.0
statsd==3.3.0
stevedore==1.30.0
testtools==2.3.0
traceback2==1.4.0
unittest2==1.1.0
urllib3==1.24.1
vine==1.1.4
WebOb==1.8.4
wrapt==1.10.11
